Implement dataset operations
-> Introduced Dataset entity
-> Introduced DatasetController that contains endpoints for Dataset operations
-> Added Multipart configuration to application.yml
-> Introduced DatasetService to handle business logic
-> Introduced Dataset repository
-> Added nessesery dependencies to pom.xml
-> Introduced FileValidator to restrict acceptable file types to csv and xls,xlsx

Notes:
- User can upload, download, delete and retrieve available dataset files.
- Users can either upload in a publicly accessed directory or a private one
- files on the filesystem are mapped to the Dataset Entity
- The combination of the following keys [filename, access, user_id] ensures that a file is unique
  (filesystem directory tree is constructed accordingly)

+++ hcvision-server/src/main/java/com/hcvision/hcvisionserver/user/User.java
@@ -1,5 +1,6 @@
 package com.hcvision.hcvisionserver.user;
 
+import com.hcvision.hcvisionserver.user.dto.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,7 +21,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue()
-    private Integer id;
+    private Long id;
     private String firstName;
     private String lastName;
     private String email;
